Refactor DetectedFrame and update references.
Reorganized the `DetectedFrame` class by moving it to `domain.frame`. Updated all references across the codebase to reflect the new import path and field renaming (`frame_number` to `no`). This improves clarity and aligns with project structure conventions.
