+++ OTVision/application/detect/detected_frame_producer.py
@@ -2,8 +2,7 @@ from typing import Generator
 
 from OTVision.abstraction.pipes_and_filter import Filter
 from OTVision.domain.detect_producer_consumer import DetectedFrameProducer
-from OTVision.domain.detection import DetectedFrame
-from OTVision.domain.frame import Frame
+from OTVision.domain.frame import DetectedFrame, Frame
 from OTVision.domain.input_source_detect import InputSourceDetect
