+++ OTVision/domain/detection.py
@@ -1,5 +1,4 @@
 from dataclasses import dataclass
-from datetime import datetime
 
 
 @dataclass(frozen=True, repr=True)
@@ -10,11 +9,3 @@ class Detection:
     y: float
     w: float
     h: float
-
-
-@dataclass(frozen=True)
-class DetectedFrame:
-    source: str
-    frame_number: int
-    occurrence: datetime
-    detections: list[Detection]
