+++ tests/application/detect/test_detected_frame_producer.py
@@ -4,8 +4,7 @@ from OTVision.abstraction.pipes_and_filter import Filter
 from OTVision.application.detect.detected_frame_producer import (
     SimpleDetectedFrameProducer,
 )
-from OTVision.domain.detection import DetectedFrame
-from OTVision.domain.frame import Frame
+from OTVision.domain.frame import DetectedFrame, Frame
 from OTVision.domain.input_source_detect import InputSourceDetect
 
 INPUT_SOURCE_GENERATOR = Mock()
